+++ Modules/NVMA/NVMA.h
@@ -54,11 +54,11 @@ void NVMA_Get_LR_TX_SF(uint8_t *sf);
 void NVMA_Set_LR_RX_SF(uint8_t sf);
 void NVMA_Get_LR_RX_SF(uint8_t *sf);
 
-void NVMA_Set_LR_TX_BW(uint32_t bw);
-void NVMA_Get_LR_TX_BW(uint32_t *bw);
+void NVMA_Set_LR_TX_BW(uint8_t bw);
+void NVMA_Get_LR_TX_BW(uint8_t *bw);
 
-void NVMA_Set_LR_RX_BW(uint32_t bw);
-void NVMA_Get_LR_RX_BW(uint32_t *bw);
+void NVMA_Set_LR_RX_BW(uint8_t bw);
+void NVMA_Get_LR_RX_BW(uint8_t *bw);
 
 void NVMA_Set_LR_TX_IQ(uint8_t iq);
 void NVMA_Get_LR_TX_IQ(uint8_t *iq);
